05-integration-hubs/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
prometheus-client==0.19.0
pyjwt==2.8.0
python-multipart==0.0.6
